+++ test/objects.spec.js
@@ -689,4 +689,31 @@ describe('Object Functions', () => {
       b: 42,
     })
   })
+  it('commonKeys', () => {
+    let providers = {
+      mongo: {},
+      elasticsearch: {}
+    }
+    let schema = {
+      fields: {},
+      elasticsearch: {},
+      mongo: {},
+    }
+    
+    expect(F.commonKeys(providers, schema)).to.deep.equal(['elasticsearch', 'mongo'])
+    expect(F.commonKeys(providers)(schema)).to.deep.equal(['elasticsearch', 'mongo'])
+  })
+  it('firstCommonKey', () => {
+    let providers = {
+      mongo: {},
+      elasticsearch: {}
+    }
+    let schema = {
+      fields: {},
+      elasticsearch: {},
+      mongo: {},
+    }
+    
+    expect(F.firstCommonKey(providers, schema)).to.equal('elasticsearch')
+  })
 })