+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add `walkAsync`
 - Add `findIndexedAsync` (used internally and not documented, but exported for testing)
 - Add `mapArgs`
+- Add `commonKeys` and `firstCommonKey`
 
 # 1.65.0
 - Add support for regular expression arguments to `highlight`